fix: changed national id data type

+++ Models/Person.cs
@@ -9,7 +9,7 @@ public class Person {
     public string? Id {get; set;}
     public string FirstName {get; set;} = default!;
     public string LastName {get; set;} = default!;
-    public int NationalID {get; set;} = default!;
+    public ulong NationalID {get; set;} = default!;
     public int Age {get; set;}
     public string? AgeStage {get; set;}
 }
